toolkit/components/passwordmgr/test/mochitest/test_autocomplete_basic_form_subdomain.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test that logins with non-exact match origin appear in autocomplete dropdown</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: logins with non-exact match origin appear in autocomplete dropdown

<script>
let origin = window.location.origin;
let lastDot = origin.lastIndexOf(".");
let suffix = origin.slice(lastDot);

let baseHost = "http://example" + suffix;
let baseSecureHost = "https://example" + suffix;
let oldHost = "http://old.example" + suffix;
let oldSecureHost = "https://old.example" + suffix;
let newHost = "https://new.example" + suffix;

addLoginsInParent(
  // The first two logins should never be visible on https: versions of
  // *.example.com since the login is for http: and an https: login exists for this username.
  [oldHost, oldSecureHost, null, "dsdu1", "dsdp1new", "uname", "pword"],
  [baseHost, baseSecureHost, null, "dsdu1", "dsdp1", "uname", "pword"],
  [oldSecureHost, oldSecureHost, null, "dsdu1", "dsdp1", "uname", "pword"],
  [baseSecureHost, baseSecureHost, null, "dsdu1", "dsdp1", "uname", "pword"],
  [newHost, newHost, null, "dsdu1", "dsdp1prime", "uname", "pword"]);

var setupScript = runInParent(function setup() {
  addMessageListener("getDateString", () => {
    let dateAndTimeFormatter = new Services.intl.DateTimeFormat(undefined, { dateStyle: "medium" });
    return dateAndTimeFormatter.format(new Date());
  });
});
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 tests multiple matching logins -->
  <form id="form1" action="https://otherexample.com/formtest.js" onsubmit="return false;">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">
    <button type="submit">Submit</button>
  </form>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

const uname = getFormElementByName(1, "uname");
const pword = getFormElementByName(1, "pword");

// Restore the form to the default state.
function restoreForm() {
  uname.value = "";
  pword.value = "";
  uname.focus();
}

function spinEventLoop() {
  return Promise.resolve();
}

add_setup(async () => {
  await SpecialPowers.pushPrefEnv({
    set: [
      ["signon.includeOtherSubdomainsInLookup", true],
    ],
  });
  listenForUnexpectedPopupShown();
});

add_task(async function test_form1_initial_empty() {
  await SimpleTest.promiseFocus(window);

  // Make sure initial form is empty.
  checkLoginForm(uname, "", pword, "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

/* For this testcase, there exists two logins for this origin
 * on different subdomains but with different passwords. Both logins
 * should appear in the autocomplete popup.
 */
add_task(async function test_form1_menu_shows_two_logins_same_usernames_for_different_subdomain() {
  await SimpleTest.promiseFocus(window);
  // Trigger autocomplete popup
  restoreForm();
  const autocompleteItems = await openAutocompletePopup();

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  // The logins are added "today" and since they are duplicates, the date that they were last
  // changed will be appended.
  let dateString = await setupScript.sendQuery("getDateString");
  let username = `dsdu1 (${dateString})`;

  checkAutoCompleteResults(autocompleteItems, [username, username], window.location.host, "Check all menuitems are displayed correctly.");

  synthesizeKey("KEY_ArrowDown"); // first item
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting

  synthesizeKey("KEY_Enter");
  await promiseFormsProcessedInSameProcess();
  is(pword.value, "dsdp1", "password should match the login that was selected");
  checkLoginForm(uname, "dsdu1", pword, "dsdp1");

  restoreForm();

  await openAutocompletePopup();

  synthesizeKey("KEY_ArrowDown"); // first item
  synthesizeKey("KEY_ArrowDown"); // second item
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting

  synthesizeKey("KEY_Enter");
  await promiseFormsProcessedInSameProcess();
  is(pword.value, "dsdp1prime", "Password should match the login that was selected");
  checkLoginForm(uname, "dsdu1", pword, "dsdp1prime");
});
</script>
</pre>
</body>
</html>
